ml/requirements.txt
absl-py==0.8.1
astor==0.8.0
boto3==1.10.26
botocore==1.13.26
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
console-menu==0.6.0
cycler==0.10.0
docutils==0.15.2
enum-compat==0.0.3
Flask==1.1.1
future==0.18.2
gast==0.2.2
gluoncv==0.5.0
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.8.4
grpcio==1.25.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
iterfzf==0.4.0.17.3
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mxnet-mkl==1.5.1.post0
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.3
Pillow==6.2.1
protobuf==3.10.0
protobuf3-to-dict==0.1.5
psutil==5.6.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.2.1
sagemaker==1.44.1
scikit-learn==0.21.3
scipy==1.3.2
seaborn==0.9.0
six==1.13.0
sklearn==0.0
sns==0.1
tabulate==0.8.6
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tqdm==4.38.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
